Deving manual reprocessing of 500Da search

+++ scripts/500Da_search.dir.bash
@@ -70,11 +70,15 @@ while [ $# -ne 0 ] ; do
 		
 		base=${file%.*}		#	drop the extension
 
-		cmd="java -Xmx16G -jar $MSGFPlus -s $file -o $base.500Da.mzid -d uniprot_reviewed_April2016.fasta -t 500Da -ti 0,1 -inst 1 -protocol 1 -mod Mods2.txt -n 10 -addFeatures 1"
-		echo $cmd
-		$cmd
-		if [ $? -ne 0 ] ; then
-			echo $cmd >> failed_commands
+		if [ ! -e $base.500Da.mzid ] ; then
+
+			cmd="java -Xmx32G -jar $MSGFPlus -s $file -o $base.500Da.mzid -d uniprot_reviewed_April2016.fasta -t 500Da -ti 0,1 -inst 1 -protocol 1 -mod Mods2.txt -n 10 -addFeatures 1"
+			echo $cmd
+#			$cmd
+#			if [ $? -ne 0 ] ; then
+#				echo $cmd >> failed_commands
+#			fi
+
 		fi
 
 #	check return status?
@@ -82,11 +86,15 @@ while [ $# -ne 0 ] ; do
 #	Do what if fails?
 
 
-		cmd="java -cp $MSGFPlus edu.ucsd.msjava.ui.MzIDToTsv -i $base.500Da.mzid -showFormula 1"
-		echo $cmd
-		$cmd
-		if [ $? -ne 0 ] ; then
-			echo $cmd >> failed_commands
+		if [ ! -e $base.500Da.tsv ] ; then
+
+			cmd="java -cp $MSGFPlus edu.ucsd.msjava.ui.MzIDToTsv -i $base.500Da.mzid -showFormula 1"
+			echo $cmd
+#			$cmd
+#			if [ $? -ne 0 ] ; then
+#				echo $cmd >> failed_commands
+#			fi
+
 		fi
